requirements.txt
customtkinter==5.2.2
ffmpeg-python==0.2.0
Pillow==10.1.0
tkinter-tooltip==2.1.0 